requirements.txt
streamlit>=1.35.0
requests>=2.32.0
python-dotenv>=1.0.1
ruff>=0.4.3
pytest>=8.1.1 